pin nix versions, & ignore nix outputs
We had to adjust some package versions to match what nix has. The
new versions seem to be compatible with cabal, stack and nix. We also
add `result` to `.gitignore`, as it's the default name of the output
from `nix-build`, and make the other entries absolute as well, as
they ought to be anyway.

We also rename `shell.nix` to `default.nix` as it seems to now be
working for building the whole project, after pinning the GHC version
and related packages etc. to ghc-9.0.2.

I am still trying to figure out how to run the test suite on demand
and show the output via nix, but it does get run when we call
`nix-build` and any failing test cases are shown.